move seed back to db
