+++ apps/main/e2e/global-setup.ts
@@ -1,8 +1,24 @@
 import { mkdir } from "node:fs/promises";
 import { request } from "@zoonk/e2e/fixtures";
-import { E2E_USERS } from "./seed/users";
 
 const BASE_URL = "http://localhost:3000";
+
+export const E2E_USERS = {
+  logout: {
+    email: "e2e-logout@zoonk.test",
+    password: "password123",
+  },
+  noProgress: {
+    email: "e2e-new@zoonk.test",
+    password: "password123",
+  },
+  withProgress: {
+    email: "e2e-progress@zoonk.test",
+    password: "password123",
+  },
+} as const;
+
+export type E2EUserKey = keyof typeof E2E_USERS;
 
 async function authenticateUser(
   name: string,
@@ -37,6 +53,3 @@ export default async function globalSetup(): Promise<void> {
     ),
   );
 }
-
-// biome-ignore lint/performance/noBarrelFile: re-exporting for convenience
-export { E2E_USERS, type E2EUserKey } from "./seed/users";
