+++ apps/main/package.json
@@ -49,7 +49,6 @@
     "dev": "next dev",
     "e2e": "dotenv -e .env.e2e -- playwright test",
     "e2e:ui": "dotenv -e .env.e2e -- playwright test --ui",
-    "seed:e2e": "dotenv -e .env.e2e -- tsx e2e/seed/index.ts",
     "start": "next start",
     "stripe:listen": "stripe listen --forward-to http://localhost:3000/api/auth/stripe/webhook",
     "test": "vitest run",
