+++ packages/db/src/prisma/seed/users.ts
@@ -2,6 +2,9 @@ import type { PrismaClient, User } from "../../generated/prisma/client";
 
 export type SeedUsers = {
   admin: User;
+  e2eLogout: User;
+  e2eNoProgress: User;
+  e2eWithProgress: User;
   logoutTest: User;
   member: User;
   owner: User;
@@ -53,5 +56,47 @@ export async function seedUsers(prisma: PrismaClient): Promise<SeedUsers> {
     where: { email: "logout-test@zoonk.test" },
   });
 
-  return { admin, logoutTest, member, owner };
+  // E2E test users with well-known credentials for auth testing
+  const e2eWithProgress = await prisma.user.upsert({
+    create: {
+      email: "e2e-progress@zoonk.test",
+      emailVerified: true,
+      name: "E2E Progress User",
+      role: "member",
+    },
+    update: {},
+    where: { email: "e2e-progress@zoonk.test" },
+  });
+
+  const e2eNoProgress = await prisma.user.upsert({
+    create: {
+      email: "e2e-new@zoonk.test",
+      emailVerified: true,
+      name: "E2E New User",
+      role: "member",
+    },
+    update: {},
+    where: { email: "e2e-new@zoonk.test" },
+  });
+
+  const e2eLogout = await prisma.user.upsert({
+    create: {
+      email: "e2e-logout@zoonk.test",
+      emailVerified: true,
+      name: "E2E Logout User",
+      role: "member",
+    },
+    update: {},
+    where: { email: "e2e-logout@zoonk.test" },
+  });
+
+  return {
+    admin,
+    e2eLogout,
+    e2eNoProgress,
+    e2eWithProgress,
+    logoutTest,
+    member,
+    owner,
+  };
 }
